nvme: add windows support for rpmb hash functions
Moved hash functions used in RPMB into new util/hash.h
with separate linux and windows implementations
