+++ util/meson.build
@@ -10,7 +10,8 @@ util_sources = [
 
 if host_system == 'windows'
     util_sources += [
-        'util/sighdl-win.c'
+        'util/sighdl-win.c',
+        'util/hash-win.c',
     ]
 else
     util_sources += [
@@ -18,6 +19,7 @@ else
         'util/utils.c',
         'util/table.c',
         'util/dashboard.c',
+        'util/hash-linux.c',
     ]
 endif
